% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{UpdateVCF}
\alias{UpdateVCF}
\title{Update genotypes of vcf file}
\usage{
UpdateVCF(haps, pref, n, vcf, compress = FALSE)
}
\arguments{
\item{haps}{character vector with haplotypes}

\item{pref}{path/to/pref, where `pref` is a basename of output file. vcf
extention is added.}

\item{n}{ordinal number to mark updated files automatically. If provided,
\emph{m} followed by this number is added to the basename of the file.}

\item{vcf}{vcf-class object loaded with \emph{readVcf()} function from
VariantAnnotation package. The genotypes of this object we want to update.}

\item{compress}{logical. If TRUE, the output is written bgzip-compressed
with the .vcf.gz extention. Such files are ready for tabix indexing.
The default is FALSE (plain .vcf as in previous versions).}
}
\description{
Update GT fields of vcf file by haplotypes provided. Updated file is saved
under the name, which can be numbered to distinguish files in a set. It is
useful, when you create a set of vcf files where GT fields are masked in
different ways. These files will be further imputed. The updated genotypes
are written as unphased.
}
