% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastphase.R
\name{WriteFastPHASE}
\alias{WriteFastPHASE}
\title{Save a vector with haplotypes as a simplified fastPHASE file ready for
imputation.}
\usage{
WriteFastPHASE(g, pref, n = NULL)
}
\arguments{
\item{g}{character vector with haplotypes}

\item{pref}{basename of output file. inp extention is added.}

\item{n}{ordinal number of masks. If provided, 'm' followed by this number
will be added to the basename of masked files.}
}
\value{
No value
}
\description{
Save a vector with haplotypes as a simplified fastPHASE file ready for
imputation.
}
