% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_masks.R
\name{ApplyMasks}
\alias{ApplyMasks}
\title{Apply set of masks to sequencies.}
\usage{
ApplyMasks(g, masks, pref)
}
\arguments{
\item{g}{Character vector with sequences}

\item{masks}{List of masks as binary matrices}

\item{pref}{path/to/prefix to save the result. The number of files created
equals to the length of masks variable. The filenames are generated
automatically like this: \code{prefix.m{n}.inp}, where \code{prefix} is a user defined
string, \code{n} is an ordinal number of the mask.}
}
\value{
No values
}
\description{
Function applies a set of masks to sequencies. The output is saved as
fastPHASE simpilfied *.inp files. Simplified means that it doesn't include
sample ids.
}
