% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quality.R
\name{EstimateQuality}
\alias{EstimateQuality}
\title{Estimate the quality of imputation}
\usage{
EstimateQuality(origin, masks, imputed, id = NULL)
}
\arguments{
\item{origin}{path to original unmasked file}

\item{masks}{path/to/masks.RDS, where \code{masks.RDS} keeps the masks
created upstream with \code{\link{GenerateMaskSet}} and saved as
\code{*.RDS}.}

\item{imputed}{vector of paths to files imputed. In the case of fastPHASE
program the elements are \code{path/to/*_genotypes.out}. In the case of
BEAGLE - \code{path/to/*.vcf} or \code{path/to/*.vcf.gz}. The order of
elements of vector and masks should coincide.}

\item{id}{character or numeric, id of computational experiment. In case you
run several calculations with different model parameter to find the best
one, you can mark each run with id for the convinience of further
visualization. The argument is optional.}
}
\value{
A data frame with three columns: "alleles", "genotypes", and "id" if
  provided. The first two contains the errors, the third one - id of the
  experiment. The error is counted as \code{(1 - accuracy)}, where
  \code{accuracy} is a proportion of correctly imputed genotypes and alleles.
  By correctly imputed genotype we mean that both alleles coincided with the
  original ones. The function returns the values for one set of test files.
}
\description{
Counts the discordance of genotypes imputed.
}
