% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quality.R
\name{EstimateQuality}
\alias{EstimateQuality}
\title{Estimate the quality of imputation}
\usage{
EstimateQuality(origin, masks, imputed, id = NULL)
}
\arguments{
\item{origin}{path to original unmasked file}

\item{masks}{path/to/masks.RDS, where \emph{masks.RDS} keeps the masks
created upstream with \emph{\link{GenerateMaskSet}} and saved as
\emph{*.RDS}.}

\item{imputed}{vector of paths to files imputed. In the case of fastPHASE,
the elements are \emph{path/to/*_genotypes.out}. In the case of vcf files,
the elements are \emph{path/to/*.vcf} or \emph{path/to/*.vcf.gz}. The order
of elements of vector and masks should coincide.}

\item{id}{character or numeric, id of computational experiment. If you run
several calculations with different model parameter to find the optimal
one, you can mark each run with id for the convinience of further
visualization. The argument is optional.}
}
\value{
A data frame with two columns: "discordance" and "id". If \emph{id}
  argument is not provided, only "discordance" is returned. It contains a
  proportion of wrognly imputed genotypes. By this we mean that one or two
  alleles are not coincided with the original ones. The function returns the
  values for one set of test files.
}
\description{
Counts the discordance of genotypes imputed.
}
