% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mask_set.R
\name{GenerateMaskSet}
\alias{GenerateMaskSet}
\title{Generate a set of masks}
\usage{
GenerateMaskSet(g, n, p)
}
\arguments{
\item{g}{Character vector with sequences}

\item{n}{Number of masks to be generated}

\item{p}{Proportion of genotypes to be masked at each loci}
}
\value{
A list of length \emph{n} containing masks as matrices
}
\description{
Given a proportion of genotypes to be masked at each loci
\emph{GenerateMaskSet} samples a set of different masks.
The number of masks in a set can vary.
}
