% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quality.R
\name{PlotDiscordance}
\alias{PlotDiscordance}
\title{Boxpolt discordance}
\usage{
PlotDiscordance(fname, tl = NULL, stl = NULL, id = NULL, ...)
}
\arguments{
\item{fname}{path to dataframe with two columns 'disc' and 'id'}

\item{tl}{optional title of the plot}

\item{stl}{optional subtitle of the plot}
}
\value{
ggplot object
}
\description{
Draw boxplots of discordance estimated for different model parameters.
}
