% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastphase.R
\name{ReadFastPHASE}
\alias{ReadFastPHASE}
\title{Read fastPHASE files}
\usage{
ReadFastPHASE(inp)
}
\arguments{
\item{inp}{path/to/filename.inp}
}
\value{
A character vector with strings representing the genotypes. Two
 sequencies per individual.
}
\description{
\code{ReadfastPHASE} parses fastPHASE *.inp files. There are two types of
them: with ids of samples provided or not. The function loads the second one,
simplified. The alleles can be coded as letters or numbers, but missing ones
should be as "?".
}
