% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastphase.R
\name{ReadFastPHASE}
\alias{ReadFastPHASE}
\title{Read fastPHASE files}
\usage{
ReadFastPHASE(inp)
}
\arguments{
\item{inp}{path/to/filename.inp}
}
\value{
A character vector with strings of letters representing haplotypes.
 The length of vector is 2N, where N is the number of diploid individuals.
}
\description{
Parses fastPHASE *.inp files. There are two types of them: with ids of samples
provided or not. The function loads both types. The alleles can be coded as
letters or numbers, but missing ones should be "?".
}
