% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{ReadVCF}
\alias{ReadVCF}
\title{Read vcf files}
\usage{
ReadVCF(x, ...)
}
\arguments{
\item{x}{path/to/filename.vcf or vcf.gz}

\item{...}{an optional parameter \emph{swap} that can be passed to
\emph{Geno2Haps()} function. If \emph{swap} is TRUE, the haplotypes of an
individual are swapped.}
}
\value{
A list with two elements: vcf-class object and a character vector which
  elements are strings of letters A, T, C, G, and symbol '?'. The length of
  vector is 2N, where N is the number of diploid individuals.
  The length of the string equals to the number of markers in vcf file.
}
\description{
Read vcf and vcf.gz files having both phased or unphased genotypes. The
missing genotypes in vcf file can be presented as '.' or './.'
\emph{ReadVCF()} is based on \emph{readVcf()} function from
VariantAnnotation R package that loads a vcf file as vcf-class object.
\emph{ReadVCF()} extracts the genotypes from vcf-class object and converts them
in a vector of strings formed by letters A, T, C, G and a symbol of missing data '?'.
}
